+++ localization/landmark_based_localizer/ar_tag_based_localizer/src/ar_tag_based_localizer.cpp
@@ -42,7 +42,7 @@
  or implied, of Rafael Muñoz Salinas.
  ********************************/
 
-#include "ar_tag_based_localizer/ar_tag_based_localizer_core.hpp"
+#include "ar_tag_based_localizer.hpp"
 
 #include <Eigen/Core>
 #include <Eigen/Geometry>
